requirements.txt
datasets==2.20.0
numpy<2.0.0
packaging==24.1
pandas==2.2.2
Pillow==10.4.0
scikit_learn==1.4.2
torch>=2.4.0
torchmetrics==1.4.0
torchsummary
torchvision==0.19.0
tqdm==4.66.4
transformers==4.42.4
tensorboard==2.17.0
